combine abstract_spinnake_common and spinkker
